requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.10.0
joblib==1.0.1
numpy==1.20.3
Pillow==8.1.0
pytz==2020.1
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0